requirements.txt
websocket-client>=1.7.0
requests>=2.31.0
PyYAML>=6.0
